requirements.txt
Flask==3.1.0
line-bot-sdk==3.17.1
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.1
gunicorn==21.2.0